requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.13.0=pyhd8ed1ab_0
astor=0.8.1=pyh9f0ad1d_0
asttokens=2.0.5=pypi_0
attrs=21.2.0=pypi_0
automat=20.2.0=pypi_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
brotlipy=0.7.0=py36h68aa20f_1001
ca-certificates=2021.10.8=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py36ha15d459_0
cffi=1.14.5=py36he58ceb7_0
chardet=4.0.0=py36ha15d459_1
click=7.1.2=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
constantly=15.1.0=pypi_0
cryptography=3.4.7=py36hd0de82c_0
cudatoolkit=10.0.130=h5327add_8
cycler=0.10.0=py_2
dataclasses=0.8=pyh787bdff_0
decorator=5.0.9=pyhd8ed1ab_0
dlib=19.21.1=py36he9698b0_0
docopt=0.6.2=py_1
docutils=0.17.1=pypi_0
editorconfig=0.12.3=pypi_0
einops=0.3.2=pyhd8ed1ab_0
et_xmlfile=1.0.1=py_1001
fire=0.4.0=pyh44b312d_0
freeglut=3.2.1=h0e60522_2
freetype=2.10.4=h546665d_1
future=0.18.2=pypi_0
gast=0.4.0=pyh9f0ad1d_0
google-auth=1.35.0=pypi_0
google-pasta=0.2.0=pyh8c360ce_0
grad-cam=1.0.0=pypi_0
graphviz=2.38=hfd603c8_2
grpcio=1.38.0=py36h4374274_0
h5py=3.1.0=nompi_py36hf359dfe_100
hdf5=1.10.6=nompi_h5268f04_1114
hyperlink=21.0.0=pypi_0
icu=68.1=h0e60522_0
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=pypi_0
importlib-metadata=4.5.0=py36ha15d459_0
imutils=0.5.4=py36ha15d459_0
incremental=21.3.0=pypi_0
intel-openmp=2021.2.0=h57928b3_616
ipdb=0.13.9=pyhd8ed1ab_0
ipykernel=4.6.1=py36_0
ipython=5.8.0=py36_1
ipython_genutils=0.2.0=py_1
jasper=2.0.14=h77af90b_2
jbig=2.1=h8d14728_2003
jdcal=1.4.1=py_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
jsonpatch=1.32=pypi_0
jsonpointer=2.1=pypi_0
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py36ha15d459_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py36he95197e_1
krb5=1.19.1=hbae68bd_0
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libclang=11.1.0=default_h5c34c98_1
libcurl=7.77.0=h789b8ee_0
libdeflate=1.7=h8ffe710_5
liblapack=3.9.0=9_mkl
liblapacke=3.9.0=9_mkl
libopencv=4.5.0=py36ha8f7756_7
libpng=1.6.37=h1d00b33_2
libprotobuf=3.17.2=h7755175_0
libsodium=1.0.18=h8d14728_1
libssh2=1.9.0=h680486a_6
libtiff=4.3.0=h0c97f57_1
libwebp-base=1.2.0=h8ffe710_2
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pyhd8ed1ab_0
matplotlib-base=3.3.3=py36h1abdf75_0
mediapipe=0.8.3=pypi_0
mkl=2021.2.0=hb70f87d_389
mpmath=1.2.1=pypi_0
msys2-conda-epoch=20160418=1
ninja=1.10.2=h5362a0b_0
numpy=1.18.5=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.0=py36ha15d459_7
opencv-python=4.5.3.56=pypi_0
openjpeg=2.4.0=hb211442_1
openpyxl=3.0.8=pyhd8ed1ab_0
openssl=1.1.1l=h8ffe710_0
paho-mqtt=1.5.1=pypi_0
pandas=1.1.5=py36he38d939_0
path=16.2.0=pypi_0
patsy=0.5.1=py_0
pickleshare=0.7.5=py_1003
pillow=8.2.0=py36h9cbe6be_1
pip=21.2.2=pypi_0
pipreqs=0.4.10=py_0
prompt_toolkit=1.0.15=py_1
protobuf=3.17.2=py36he2d232f_0
py-opencv=4.5.0=py36h7b2dad6_7
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.9.0=pyhd8ed1ab_0
pyhamcrest=2.0.2=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pypiwin32=223=pypi_0
pyqt5=5.15.4=pypi_0
pyqt5-plugins=5.15.4.2.2=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.0=pypi_0
pyqt5-tools=5.15.4.3.2=pypi_0
pyreadline=2.1=py36ha15d459_1003
pyserial=3.5=pyhd8ed1ab_0
pysocks=1.7.1=py36ha15d459_3
pytesseract=0.3.7=pyhd3deb0d_1
python=3.6.12=h39d44d4_0_cpython
python-dateutil=2.8.1=py_0
python-dotenv=0.18.0=pypi_0
python-graphviz=0.16=pypi_0
python_abi=3.6=1_cp36m
pytorch=1.2.0=py3.6_cuda100_cudnn7_1
pytz=2021.1=pyhd8ed1ab_0
pywin32=301=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.1.0=py36h1d5d788_0
qt=5.12.9=h5909a2a_4
qt5-applications=5.15.2.2.2=pypi_0
qt5-tools=5.15.2.1.2=pypi_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.1=py36h63cef00_0
scipy=1.5.3=py36h7ff6e69_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
send2trash=1.5.0=pypi_0
setuptools=49.6.0=py36ha15d459_3
simplegeneric=0.8.1=py_1
six=1.16.0=pyh6c4a22f_0
sqlite=3.35.5=h8ffe710_0
statsmodels=0.12.2=py36h14b15b0_0
sympy=1.8=pypi_0
tbb=2021.2.0=h2d74725_0
tensorboard=1.14.0=py36_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.2=pyhd8ed1ab_0
tensorflow=1.14.0=h39a9291_0
tensorflow-base=1.14.0=py36h9f0ad1d_0
tensorflow-estimator=1.14.0=py36h5ca1d4c_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h8ffe710_1
tnt=126=h57928b3_0
toml=0.10.2=pypi_0
torchfile=0.1.0=py_0
torchvision=0.4.0=py36_cu100
tornado=6.1=py36h68aa20f_1
tqdm=4.62.0=pypi_0
traitlets=4.3.3=py36h9f0ad1d_1
twisted=20.3.0=pypi_0
typing_extensions=3.10.0.0=pyha770c72_0
urllib3=1.26.5=pyhd8ed1ab_0
vc=14.1=h21ff451_1
visdom=0.1.8.9=0
vs2015_runtime=14.28.29325=h5e1d092_4
vs2017_runtime=15.4.27004.2010=1
wcwidth=0.2.5=pyh9f0ad1d_2
websocket-client=0.57.0=py36ha15d459_4
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
win_inet_pton=1.1.0=py36ha15d459_2
wincertstore=0.2=py36ha15d459_1006
wrapt=1.12.1=py36h68aa20f_3
xlrd=1.2.0=pyh9f0ad1d_1
xlsxwriter=3.0.1=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_1
yarg=0.1.9=py_1
zeromq=4.3.4=h0e60522_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_1010
zope-interface=5.4.0=pypi_0
zstd=1.5.0=h6255e5f_0
